Transfer TrustRegion and Adaptive into strategy.py in the optim directory
